% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{citxt}
\alias{citxt}
\title{format CI}
\usage{
citxt(
  est,
  low,
  high,
  d = 2,
  between = "-",
  space1 = TRUE,
  space2 = TRUE,
  exponent = FALSE
)
}
\arguments{
\item{est}{estimate}

\item{low}{lower bound}

\item{high}{upper bound}

\item{d}{digits}

\item{space1}{logical; space between 'estimate' and parenthesis}

\item{space2}{logical; space between comma and 'upper bound'}

\item{exponent}{logical; exponentiate the numbers?}
}
\description{
Format CI as text "estimate (lower bound, upper bound)". Note to self: this
is mainly a function since you (the author) always forget.
}
\examples{
e  = c(1,     1.501, 1.499)
lo  = c(0.899, 1.250, 1.000)
hi = c(1.455, 1.950, 2)
citxt(e, lo, hi)
}
