% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_tex}
\alias{clean_tex}
\title{clean LaTeX files}
\usage{
clean_tex()
}
\description{
Remove unnecessary LaTeX files. Prompt will be given before
    erasing.
}
\details{
Sometimes LaTeX leaves a mess. View source code to see what file
    endings are matched for removal.
}
