% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach-stuff.R
\name{csv_attach}
\alias{csv_attach}
\title{attach csv}
\usage{
csv_attach(cap, tab, file = NULL, dir = "table")
}
\arguments{
\item{cap}{the caption}

\item{tab}{the table to be written as csv (using \code{utils::write.cvs2})}

\item{file}{name of file; path and extension will be added automatically.}

\item{dir}{subfolder; default 'table'}
}
\description{
A function that (1) writes a csv to file, and (2) returns the
    caption with the appropriate LaTeX code to attach that file at the end.
    Currently this function will use \code{write.csv2}, since I find that it
    readable for everyone (so far).
}
