% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.r
\name{dm_on}
\alias{dm_on}
\alias{dm_off}
\alias{dm_}
\title{activate DM}
\usage{
dm_on()

dm_off()

dm_(value = NULL)
}
\arguments{
\item{value}{value for 'dm_active'}
}
\description{
set hproj option 'dm_activate' to TRUE
}
\section{Functions}{
\itemize{
\item \code{dm_off()}: inactivate DM

\item \code{dm_()}: set activation status for DM

}}
