% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_struc_functions.r
\name{doc_struc}
\alias{doc_struc}
\alias{doc_struc_dm}
\alias{chunks_info}
\alias{sections_info}
\title{document structure}
\usage{
doc_struc(file = NULL, dm = NULL)

doc_struc_dm()

chunks_info(file = NULL, all = FALSE)

sections_info(file = NULL)
}
\arguments{
\item{file}{file of interest}

\item{dm}{logical; if TRUE divert to DM-file if 'file' is NULL}

\item{all}{if TRUE some extra chunk info will be given}
}
\value{
A print out

A data frame with variables
\itemize{
   \item name: name of chunk (if any)
   \item row: start row
   \item n.row: number of rows
   \item stop: end line of chunk
   \item eval.arg: (if \code{all = TRUE}) if there is an argument specified for \code{eval}
   \item code:  (if \code{all = TRUE}) the code in the chunk
}

A data frame with variables
\itemize{
   \item name: name of section
   \item row: row number where section starts
   \item sub: number of 'sub', i.e. 0 for section, 1 for subsection,
   and 2 for subsubsection
}
}
\description{
Get information about the structure in a rnw file. Leave empty
    all arguments empty and hproj will try to figure out what you want.
}
\section{Functions}{
\itemize{
\item \code{doc_struc_dm}: shorthand for \code{doc_struc(dm = TRUE)}

\item \code{chunks_info}: get information about chunks in a rnw file

\item \code{sections_info}: get information about sections in a rnw file
}}

