% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_nchar.R
\name{equal_nchar}
\alias{equal_nchar}
\title{elements of uniformal 'nchar' length}
\usage{
equal_nchar(x, n = NULL, fill = " ", right = FALSE)
}
\arguments{
\item{x}{vector to fix}

\item{n}{uniformal nchar length for elements of x}

\item{fill}{character to fill the void}

\item{right}{fill to the right?}
}
\value{
character string
}
\description{
make nchar have the same value for all elements of x
}
\examples{
equal_nchar(x = c("A", "ABC"))
equal_nchar(x = c("A", "ABC"), 2, fill = "_")
equal_nchar(x = c("A", "ABC"), fill = ".", right = FALSE)
equal_nchar(8:12, fill = "0")
}
