% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{farben}
\alias{farben}
\title{some palettes}
\usage{
farben(p, show = FALSE)
}
\arguments{
\item{p}{character, one of the available palettes}

\item{show}{logical; if TRUE plot the palette}
}
\value{
character vector of color codes; possible a plot as a side effect
}
\description{
collection of some commonly used palettes
}
