% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.r
\name{fetch}
\alias{fetch}
\alias{fetch_object}
\alias{fetch_dots}
\title{load}
\usage{
fetch(
  ...,
  overwrite = TRUE,
  message = FALSE,
  formats = c("rdata", "rdat"),
  env = .GlobalEnv
)

fetch_object(
  name = NULL,
  overwrite = FALSE,
  message = FALSE,
  formats = c("rdata", "rdat", "Rdata", "Rdat"),
  env = .GlobalEnv
)

fetch_dots()
}
\arguments{
\item{...}{(unquoted) names of objects}

\item{overwrite}{if variable already exists in global workspace, should it be overwritten?}

\item{message}{do you want an explanatory message?}

\item{formats}{formats to look for. Default \code{c('.rdata', '.rdat')}.}

\item{env}{the environment to load into. Defaults to the global enviroment.}

\item{name}{character; the name of a variable. If missing, available objects will be listed}
}
\description{
Load an object from 'calc' directory. Use \code{fetch()} to see
    what objects are available. Objects stored with \code{keep} will also
    have meta-data recorded in 'calc/.meta'.
}
\section{Functions}{
\itemize{
\item \code{fetch_object()}: the function that does the work

\item \code{fetch_dots()}: a shorthand for loading all variables with variable names
beginning with '.'

}}
