% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach-stuff.R
\name{fig_attach}
\alias{fig_attach}
\title{attach figure}
\usage{
fig_attach(cap, ext = NULL)
}
\arguments{
\item{cap}{the caption}

\item{ext}{type of figure you want (default: 'pdf'). Note: in general, this
output must be created by knitr, so should be one of the devices given
as chunk option 'dev'.}
}
\description{
A function that will return the caption with the appropriate
    LaTeX code for attaching, meant for use with knitr:s system for creating
    figures.
}
