% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_name.r
\name{file_name}
\alias{file_name}
\title{decompose filename}
\usage{
file_name(filename)
}
\arguments{
\item{filename}{the vector of names you want decomposed}
}
\value{
A dataframe with variables 'name' and 'extension'
}
\description{
Decompose a filename into name and extension
}
\examples{
some_names <- c("foo.txt", "foo.bar.r", ".emacs", "DESCRIPTION")
file_name(some_names)
}
