% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{group_frame}
\alias{group_frame}
\title{group data.frame}
\usage{
group_frame(object, group, onto, ontoprefix = "    ")
}
\arguments{
\item{object}{a data.frame}

\item{group}{character, name of grouping variable}

\item{onto}{character/logical, the name of varible that grouping values are
entered onto, or, if TRUE, enter grouping values onto all}

\item{ontoprefix}{if not null, paste this as a prefix on the onto variable
(only applicable if 'onto' is character)}
}
\description{
split a data frame by inserting lines according to grouping specified by a
variable 'group'.
}
\examples{
d <- data.frame(hey = sample(c("John", "Jane", "Anaximandros"), 10, TRUE),
                numy = round(runif(10, 15,25), 1),
                foo = rep(LETTERS[1:3], c(2, 5, 3)),
                bar = letters[1:10],
                measx = round(rnorm(10,10,1), 1))
group_frame(object = d, group = "foo", onto = "bar")
group_frame(object = d, group = "foo", onto = TRUE)
}
