% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_marker.R
\name{id_marker}
\alias{id_marker}
\alias{sub_key}
\title{create 'id mark'}
\usage{
id_marker(
  x = "Created from <path> by <user> at <time>.",
  ...,
  file = NULL,
  chunk = NULL
)

sub_key(x, key, g = FALSE, fixed = TRUE)
}
\arguments{
\item{x}{character; string where cartain marks of the form '\if{html}{\out{<text>}}' will be
replaced}

\item{...}{things passed to \code{sub_key}}

\item{file}{character; if '\if{html}{\out{<file>}}' is used in x, this is the replacement}

\item{chunk}{character; if '\if{html}{\out{<chunk>}}' is used, this is the replacement. Note:
if \code{id_marker} is being called at runtime (by knitr), then the
actual label of the current chunk will be used.}

\item{g}{logical; use gsub instead of sub}

\item{fixed}{logical; passed to sub/gsub}
}
\value{
A string of length 1
}
\description{
Create a string that can capture user, time of creation, file, working
directory, etc.
}
\details{
Currently the following marks are replaced
\itemize{
\item \verb{<user>} by \code{Sys.getenv("USERNAME")}
\item \verb{<wd>} by \code{getwd()}
\item \verb{<time>} by \code{format(Sys.time(), format = "\%Y-\%m-\%d \%H:\%M")}
\item \verb{<file>} by 'file' (has to be supplied by the arguments)
\item \verb{<path>} by \code{file.path(wd, file)}
\item \verb{<chunk>} by 'chunk' (or, the actual chunk label if function called at
knitr runtime)
}
}
\section{Functions}{
\itemize{
\item \code{sub_key()}: substitute by key; the names of key will be replaced
by the values of key in x

}}
