% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.r
\name{keep}
\alias{keep}
\alias{keep_object}
\title{save}
\usage{
keep(...)

keep_object(name)
}
\arguments{
\item{...}{unquoted names of objects}

\item{name}{character vector; the name of one or more variable}
}
\description{
Save object(s) to 'calc' folder. Retrieve later with
    \code{fetch}
}
\details{
Some meta-data on objects saved in stored in 'calc/.meta'. That
    information can be retrieved with \code{fetch()}.
}
\section{Functions}{
\itemize{
\item \code{keep_object}: the function that does the work
}}

