% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach-stuff.R
\name{knit_proc}
\alias{knit_proc}
\alias{knit_runtime}
\title{chunk options at runtime}
\usage{
knit_proc(s)

knit_runtime()
}
\arguments{
\item{s}{the option you want}
}
\description{
A wrapper for \code{knitr::opts_current$get()}. Sometimes it is
useful to have access to current chunk options at runtime, e.g. 'label'.
}
\section{Functions}{
\itemize{
\item \code{knit_runtime()}: check if knitr is running

}}
