% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-stuff.R
\name{latex_clean}
\alias{latex_clean}
\title{clean LaTeX files}
\usage{
latex_clean()
}
\description{
Remove unnecessary LaTeX files. Prompt will be given before
    erasing.
}
\details{
Sometimes LaTeX leaves a mess. View source code to see what file
    endings are matched for removal.
}
