% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_install.r
\name{local_install}
\alias{local_install}
\title{install a local package}
\usage{
local_install(path)
}
\arguments{
\item{path}{path to local package}
}
\description{
Because I can NEVER remember what parameter choice I need
to make to make \code{install.packages} work.
}
