% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.r
\name{look}
\alias{look}
\alias{look_dm}
\title{view pdf}
\usage{
look(file = NULL, dm = NULL)

look_dm()
}
\arguments{
\item{file}{file of interest}

\item{dm}{logical; if TRUE divert to DM-file if 'file' is NULL}
}
\description{
Leave arguments empty to open the current pdf version of the
    report/DM file, or point to any pdf, as this is just a wrapper for
    \code{shell.exec}
}
\section{Functions}{
\itemize{
\item \code{look_dm}: shorthand for \code{look(dm = TRUE)}
}}

