% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.r
\name{opts_hproj}
\alias{opts_hproj}
\title{hproj options}
\usage{
opts_hproj
}
\description{
This list tries to mimic the behaviour of opts_chunk from knitr.
Currently these values are maintained with the functions in (the list)
\code{opts_hproj}: \itemize{

\item{source_file - name derived from project name}

\item{output_file - will be like source_file but appropriate file extension
(unless set manually)}

\item{version a version number as character, e.g. "Version 1". This will
appear on the LaTeX version of the report}

\item{version_latex this is the string that will determine how the version
number appears in the title (in LaTeX produced pdf:s) and will be added
automatically}

\item{all above options also exists in a version with prefix \code{dm_}}

\item{dm_active, this will indicate what file to point to (when a
dm-file is present) for functions \code{cmp}, \code{doc_struc}, \code{look},
\code{send}. This can also be toggled with \code{dm_on()} and \code{dm_off()}}

}
}
