% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chunk.r
\name{run_chunk}
\alias{run_chunk}
\alias{rblock}
\alias{rblock_dm}
\alias{cess}
\alias{cess_dm}
\title{run chunks}
\usage{
run_chunk(
  chunk = NULL,
  file = NULL,
  dm = NULL,
  envir = .GlobalEnv,
  verbose = TRUE
)

rblock(
  chunk = NULL,
  file = NULL,
  dm = NULL,
  envir = .GlobalEnv,
  verbose = TRUE
)

rblock_dm(chunk = NULL, ...)

cess(chunk = NULL, file = NULL, dm = NULL, envir = .GlobalEnv, verbose = TRUE)

cess_dm(chunk = NULL, ...)
}
\arguments{
\item{chunk}{name or number of chunk, if \code{NULL} (default) then the first
chunk will be executed}

\item{file}{the file which has the chunks, default 'source_file'}

\item{dm}{if TRUE when file is NULL, then file will be gotten from hproj options}

\item{envir}{is \code{.GlobalEnv} by default, the environemnt in which to
evaluate chunks}

\item{verbose}{if TRUE, may print some slightly unneccessary messages}

\item{...}{arguments passed from shorthand functions to \code{run_chunk}}
}
\description{
Run a chunk in a given document
}
\section{Functions}{
\itemize{
\item \code{rblock()}: \code{rblock} is an alias for \code{run_chunk}

\item \code{rblock_dm()}: \code{rblock_dm(chunk, ...)} is shorthand for
\code{run_chunk(chunk, file = NULL, dm = TRUE, ...)}

\item \code{cess()}: \code{cess} is an obsolete alias for \code{run_chunk}

\item \code{cess_dm()}: \code{cess_dm} is an obsolete alias for
\code{rblock_dm}

}}
