% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.r
\name{send}
\alias{send}
\alias{send_dm}
\alias{send_function}
\title{mark file for delivery}
\usage{
send(git = TRUE, dm = NULL)

send_dm(git = TRUE)

send_function(output_file, version, source_file, git)
}
\arguments{
\item{git}{commit? (default \code{TRUE})}

\item{dm}{logical; send the dm? If NULL DM status will be checked}

\item{output_file}{the report pdf}

\item{version}{version number}

\item{source_file}{the report rnw}
}
\description{
date and (possibly) version stamp a report and place in 'sent' folder
}
\section{Functions}{
\itemize{
\item \code{send_dm}: shorthand for \code{send(dm = TRUE)}

\item \code{send_function}: the function that does the work
}}

