% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.r
\name{cmp}
\alias{cmp}
\alias{cmp_dm}
\alias{cmp_rnw}
\title{compile main rnw file}
\usage{
cmp(dm = NULL, ...)

cmp_dm()

cmp_rnw(input, output, ...)
}
\arguments{
\item{dm}{logical; if TRUE choose dm-source. If NULL 'dm' is set to the
hproj option 'dm_active'.}

\item{...}{arguments passed to \code{\link[knitr]{knit2pdf}}}

\item{input}{file to compile}

\item{output}{output tex file}
}
\description{
Compile source or dm-source. Leave arguments empty to let
    \code{cmp} figure out which one you want.
}
\section{Functions}{
\itemize{
\item \code{cmp_dm()}: shorthand for \code{cmp(dm=TRUE)}

\item \code{cmp_rnw()}: the function that compiles; a wrapper for
\code{knitr::knit2pdf}

}}
