% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_marker.R
\name{id_marker}
\alias{id_marker}
\alias{standard_mark}
\alias{sub_key}
\title{create 'id mark'}
\usage{
id_marker(
  x = "Created by <user> at <time> from chunk '<chunk>' in <path>.",
  ...,
  file = opts_hproj$get("source_file"),
  chunk = knitr::opts_current$get("label"),
  wd = getwd(),
  path = file.path(wd, file),
  user = Sys.getenv("USERNAME"),
  time = format(Sys.time(), format = "\%Y-\%m-\%d \%H:\%M")
)

standard_mark(
  n1 = 80,
  n2 = n1,
  newline = TRUE,
  linebreak = "\\n",
  file = opts_hproj$get("source_file"),
  chunk = NULL
)

sub_key(x, key, g = FALSE, fixed = TRUE)
}
\arguments{
\item{x}{character; string where cartain marks of the form '\if{html}{\out{<text>}}' will be
replaced}

\item{...}{things passed to \code{sub_key}}

\item{file}{character; name of file}

\item{chunk}{logical; want reference to a chunk? Derived from filename if NULL}

\item{wd}{character (working directory), replaces '\if{html}{\out{<wd>}}'}

\item{path}{character (file path), replaces '\if{html}{\out{<path>}}'}

\item{user}{character (username), replaces '\if{html}{\out{<user>}}'}

\item{time}{character (date+time), replacez '\if{html}{\out{<time>}}'}

\item{n1}{integer; max length for first part of mark text}

\item{n2}{integer; max length for second part of mark text}

\item{newline}{logical; linebreak between first and second part of mark text}

\item{linebreak}{character; code for linebreak}

\item{g}{logical; use gsub instead of sub}

\item{fixed}{logical; passed to sub/gsub}
}
\value{
A string of length 1
}
\description{
Create a string that can capture user, time of creation, file, working
directory, etc.
}
\details{
Currently the following marks are replaced
\itemize{
\item \verb{<user>} by \code{Sys.getenv("USERNAME")}
\item \verb{<wd>} by \code{getwd()}
\item \verb{<time>} by \code{format(Sys.time(), format = "\%Y-\%m-\%d \%H:\%M")}
\item \verb{<file>} by 'file' (has to be supplied by the arguments)
\item \verb{<path>} by \code{file.path(wd, file)}
\item \verb{<chunk>} by 'chunk' (or, the actual chunk label if function called at
knitr runtime)
}
}
\section{Functions}{
\itemize{
\item \code{standard_mark()}: a standard marker

\item \code{sub_key()}: substitute by key; the names of key will be replaced
by the values of key in x

}}
