% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.r
\name{new_project}
\alias{new_project}
\title{create new project}
\usage{
new_project(
  name = "new_project",
  path = NULL,
  class = "ucr",
  dm = FALSE,
  RSproj = TRUE,
  git = TRUE,
  checkpoint = FALSE,
  checkpoint.date = NULL,
  go_there = FALSE,
  org = FALSE
)
}
\arguments{
\item{name}{name of the project}

\item{path}{path to project directory (else current)}

\item{class}{class of document in source file (default: 'ucr')}

\item{dm}{should a data management file be created?}

\item{RSproj}{Start a RStudio project?}

\item{git}{should git be initialized? (also a .gitignore file will be
created)}

\item{checkpoint}{should checkpoint be used?}

\item{checkpoint.date}{date for checkpoint}

\item{go_there}{Set working directory to project directory? (default: FALSE)}

\item{org}{should an org file be created?}
}
\description{
This function sets up a project directory structure along with
some files
}
\details{
This function sets up a folder with subfolders \itemize{
\item{cache: this is used by knitr}

\item{calc: this is for storage of .rdat files; 'keep' saves to here; 'fetch'
loads from here}

\item{figure: for plots (also used by knitr)}

\item{misc-docs: for all those miscellaneous documents that do not fit anywhere}

\item{sent: this is were I store things sent to client. The function
\code{send} will attach the current date (and version number, if
applicable) to the pdf version of the report and put it in this directory.}

\item{table: for (human readable) tabulated data}
}

and creates (some of them optionally), the files
\itemize{

\item{a source file (for the report) - it's name will be derived from the
project folder name.}

\item{a dm file: for those projects where the data management is more intricate
and it is better to keep analyses separate from dm}

\item{.Rprofile which will load and point hproj to the source file(s)}

\item{'references.bib': a template for bibTeX references}

\item{a .rproj file with the project name: this is an RStudio project file,
by starting this file RStudio will set the working directory and remember
what documents you were looking at. There are settings to be made that can
be project specific}

\item{.gitignore: a file that git uses to tell which files to ignore}
}

Note also that an org-file can be created (for user of org-mode), but I've
realized I do not want org-files in each project - so I don't use this
feature.

Also, this function can setup 'checkpoint' for reproducibility.
}
